ajustes

+++ jsp-servlet-crud-filtro/src/com/web/servlet/LivroController.java
@@ -27,23 +27,28 @@ public class LivroController extends HttpServlet {
 			throws ServletException, IOException {
 		String forward = "";
 		String action = request.getParameter("action");
+		String buscarLivro = request.getParameter("buscarLivro");
 
-		if (action.equalsIgnoreCase("deletar")) {
+		if (action != null && action.equalsIgnoreCase("deletar")) {
 			int livroId = Integer.parseInt(request.getParameter("livroId"));
 			dao.removeById(livroId);
 			forward = LISTAR_LIVROS;
 			request.setAttribute("livros", dao.listarTodos());
-		} else if (action.equalsIgnoreCase("editar")) {
+		} else if (action != null && action.equalsIgnoreCase("editar")) {
 			forward = INSERIR_OU_EDITAR;
 			int codigoLivro = Integer.parseInt(request.getParameter("livroId"));
 			Livro livro = dao.consultar(codigoLivro);
 			request.setAttribute("livro", livro);
-		} else if (action.equalsIgnoreCase("listarLivros")) {
+		} else if (action != null && action.equalsIgnoreCase("listarLivros")) {
 			forward = LISTAR_LIVROS;
 			request.setAttribute("livros", dao.listarTodos());
+		} else if (buscarLivro != null) {
+			List<Livro> livros = dao.consultar(buscarLivro);
+			request.setAttribute("livros", livros);
+			forward = LISTAR_LIVROS;
 		} else {
 			forward = INSERIR_OU_EDITAR;
-		}
+		} 
 
 		RequestDispatcher view = request.getRequestDispatcher(forward);
 		view.forward(request, response);
@@ -64,9 +69,6 @@ public class LivroController extends HttpServlet {
 		if ((codigoLivro == null || codigoLivro.isEmpty()) && buscarLivro == null) {
 			dao.save(livro);
 			request.setAttribute("livros", dao.listarTodos());
-		} else if (buscarLivro != null) {
-			List<Livro> livros = dao.consultar(buscarLivro);
-			request.setAttribute("livros", livros);
 		} else {
 			livro.setCodigo(Integer.parseInt(codigoLivro));
 			dao.updateTitulo(livro);
