+++ jsp-servlet-crud/src/com/web/servlet/LivroController.java
@@ -26,23 +26,28 @@ public class LivroController extends HttpServlet {
 			throws ServletException, IOException {
 		String forward = "";
 		String action = request.getParameter("action");
+		String buscarLivro = request.getParameter("buscarLivro");
 
-		if (action.equalsIgnoreCase("deletar")) {
+		if (action != null && action.equalsIgnoreCase("deletar")) {
 			int livroId = Integer.parseInt(request.getParameter("livroId"));
 			dao.removeById(livroId);
 			forward = LISTAR_LIVROS;
 			request.setAttribute("livros", dao.listarTodos());
-		} else if (action.equalsIgnoreCase("editar")) {
+		} else if (action != null && action.equalsIgnoreCase("editar")) {
 			forward = INSERIR_OU_EDITAR;
 			int codigoLivro = Integer.parseInt(request.getParameter("livroId"));
 			Livro livro = dao.consultar(codigoLivro);
 			request.setAttribute("livro", livro);
-		} else if (action.equalsIgnoreCase("listarLivros")) {
+		} else if (action != null && action.equalsIgnoreCase("listarLivros")) {
 			forward = LISTAR_LIVROS;
 			request.setAttribute("livros", dao.listarTodos());
+		} else if (buscarLivro != null) {
+			List<Livro> livros = dao.consultar(buscarLivro);
+			request.setAttribute("livros", livros);
+			forward = LISTAR_LIVROS;
 		} else {
 			forward = INSERIR_OU_EDITAR;
-		}
+		} 
 
 		RequestDispatcher view = request.getRequestDispatcher(forward);
 		view.forward(request, response);
@@ -55,17 +60,23 @@ public class LivroController extends HttpServlet {
 		livro.setTitulo(request.getParameter("titulo"));
 		livro.setAutor(request.getParameter("autor"));
 		livro.setDescricao(request.getParameter("descricao"));
-		livro.setPreco(Double.parseDouble(request.getParameter("preco")));
+		livro.setPreco(request.getParameter("preco") != null ? Double.parseDouble(request.getParameter("preco")) : 0);
 		String codigoLivro = request.getParameter("livroCodigo");
+		String buscarLivro = request.getParameter("buscarLivro");
+		
 
-		if (codigoLivro == null || codigoLivro.isEmpty()) {
-			 dao.save(livro);
+		if ((codigoLivro == null || codigoLivro.isEmpty()) && buscarLivro == null) {
+			dao.save(livro);
+			request.setAttribute("livros", dao.listarTodos());
+		} else if (buscarLivro != null) {
+			List<Livro> livros = dao.consultar(buscarLivro);
+			request.setAttribute("livros", livros);
 		} else {
 			livro.setCodigo(Integer.parseInt(codigoLivro));
-			 dao.updateTitulo(livro);
+			dao.updateTitulo(livro);
+			request.setAttribute("livros", dao.listarTodos());
 		}
 		RequestDispatcher view = request.getRequestDispatcher(LISTAR_LIVROS);
-		request.setAttribute("livros", dao.listarTodos());
 		view.forward(request, response);
 	}
